requirements.txt
pandas
plotly
psycopg2-binary
python-dotenv
sqlmodel
statsmodels
streamlit
